Add an empty version of thunk scope
This will be used when no thunks exist yet. It's important to have this
for the realtime implementation, because pakyow currently adds
properties to this element as a placeholder until records are added.
Without an "empty version" you would have to refresh the page when the
first record is created. After that everything would work as expected.